restructure express API to be modular
