+++ api/express/package.json
@@ -3,6 +3,7 @@
     "axios": "^1.8.4",
     "express": "^4.21.2",
     "form-data": "^4.0.2",
+    "mongodb": "^6.15.0",
     "multer": "^1.4.5-lts.2"
   },
   "name": "express",
